Based on the provided git diff output, I will generate a conventional commit message.
The change is a modification to the `laravellocalization.php` configuration file, specifically changing the value of `'useAcceptLanguageHeader'` from `true` to `false`.

The primary type of this change is **chore**, as it is a maintenance task or tooling change, and it does not introduce a new feature, fix a bug, or significantly refactor the code.

The scope of this change is **config**, as it modifies a configuration file.

The description of this change should focus on the most significant change, which is the disabling of the use of the Accept Language Header.

Here is the generated commit message:

chore(config): disable use of Accept Language Header